requirements.txt
kafka-python
geopy
pandas
xgboost
scikit-learn
streamlit